+++ lib/src/adapty_version.dart
@@ -1,1 +1,1 @@
-const String adaptySDKVersion = '4.1.0';
+const String adaptySDKVersion = '4.1.1';